% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bvalue}
\alias{bvalue}
\title{B-Value Calculation}
\usage{
bvalue(small_delta, big_delta, G)
}
\arguments{
\item{small_delta}{A numeric value specifying the duration of the gradient
pulse in ms.}

\item{big_delta}{A numeric value specifying the duration between the gradient
pulses in ms.}

\item{G}{A numeric value specifying the strength of the gradient in
\eqn{\mu}T/\eqn{\mu}m.}
}
\value{
A numeric value storing the predicted b-value in
ms/\eqn{\mu}m\eqn{^2}.
}
\description{
A function to calculate the b-value for a given set of experimental
parameters.
}
\examples{
bvalue(small_delta = 30, big_delta = 30, G = 0.040)
}
