% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularly-shaped-compartments.R
\name{CallaghanCompartment}
\alias{CallaghanCompartment}
\title{Callaghan's model for restricted diffusion in a cylinder}
\description{
A class to model restricted diffusion in a cylinder using the
Callaghan's model.
}
\references{
Callaghan, P. T. (1995). Pulsed-gradient spin-echo NMR for
planar, cylindrical, and spherical pores under conditions of wall
relaxation. Journal of magnetic resonance, Series A, 113(1), 53-59.
}
\section{Super class}{
\code{\link[midi:CircularlyShapedCompartment]{midi::CircularlyShapedCompartment}} -> \code{CallaghanCompartment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CallaghanCompartment-clone}{\code{CallaghanCompartment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="midi" data-topic="CircularlyShapedCompartment" data-id="get_signal"><a href='../../midi/html/CircularlyShapedCompartment.html#method-CircularlyShapedCompartment-get_signal'><code>midi::CircularlyShapedCompartment$get_signal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="CircularlyShapedCompartment" data-id="initialize"><a href='../../midi/html/CircularlyShapedCompartment.html#method-CircularlyShapedCompartment-initialize'><code>midi::CircularlyShapedCompartment$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallaghanCompartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CallaghanCompartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallaghanCompartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
