% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-package.R
\docType{package}
\name{midi-package}
\alias{midi}
\alias{midi-package}
\title{midi: Microstructure Information from Diffusion Imaging}
\description{
A collection of diffusion models and tools to extract microstructure information from diffusion MRI data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/astamm/midi}
  \item \url{https://astamm.github.io/midi/}
  \item Report bugs at \url{https://github.com/astamm/midi/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
