% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_R6.R
\name{midi}
\alias{midi}
\title{midi class constructor}
\description{
midi class constructor

midi class constructor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{midi$new()}}
\item \href{#method-print}{\code{midi$print()}}
\item \href{#method-encode}{\code{midi$encode()}}
\item \href{#method-cut}{\code{midi$cut()}}
\item \href{#method-plot}{\code{midi$plot()}}
\item \href{#method-select}{\code{midi$select()}}
\item \href{#method-rename}{\code{midi$rename()}}
\item \href{#method-set_tempo}{\code{midi$set_tempo()}}
\item \href{#method-shift_hstep}{\code{midi$shift_hstep()}}
\item \href{#method-shift_degree}{\code{midi$shift_degree()}}
\item \href{#method-negative_harmony}{\code{midi$negative_harmony()}}
\item \href{#method-play}{\code{midi$play()}}
\item \href{#method-clone}{\code{midi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$new(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$print(..., n = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode"></a>}}
\if{latex}{\out{\hypertarget{method-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$encode(to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cut"></a>}}
\if{latex}{\out{\hypertarget{method-cut}{}}}
\subsection{Method \code{cut()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$cut(start = 0, end = Inf, unit = c("sec", "tick"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\if{latex}{\out{\hypertarget{method-select}{}}}
\subsection{Method \code{select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$select(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename"></a>}}
\if{latex}{\out{\hypertarget{method-rename}{}}}
\subsection{Method \code{rename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$rename(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_tempo"></a>}}
\if{latex}{\out{\hypertarget{method-set_tempo}{}}}
\subsection{Method \code{set_tempo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$set_tempo(
  tempo,
  metric = c("bpm", "relative", "micro_seconds_per_quarter_note")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_hstep"></a>}}
\if{latex}{\out{\hypertarget{method-shift_hstep}{}}}
\subsection{Method \code{shift_hstep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$shift_hstep(n, at = -1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_degree"></a>}}
\if{latex}{\out{\hypertarget{method-shift_degree}{}}}
\subsection{Method \code{shift_degree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$shift_degree(n, scale, at = -1, stay_home = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-negative_harmony"></a>}}
\if{latex}{\out{\hypertarget{method-negative_harmony}{}}}
\subsection{Method \code{negative_harmony()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$negative_harmony(pivot = 60, at = -1, stay_home = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-play"></a>}}
\if{latex}{\out{\hypertarget{method-play}{}}}
\subsection{Method \code{play()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$play(left_track = NULL, right_track = left_track)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
