% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_midi.R
\docType{data}
\name{parse_midi}
\alias{parse_midi}
\title{Load a midi file into a midi object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
parse_midi
}
\arguments{
\item{file}{path to midi file}
}
\description{
Load a midi file into a midi object
}
\keyword{datasets}
