% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder-bundle-geometry.R
\name{plot3d.bundle}
\alias{plot3d.bundle}
\title{Plots a 3D representation of a cylinder bundle using \code{rgl}}
\usage{
\method{plot3d}{bundle}(x, widget = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{bundle} as generated by
\code{\link[=simulate_bundle]{simulate_bundle()}}.}

\item{widget}{A logical value indicating whether to return an HTML widget for
reporting. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to be passed to the \code{\link[rgl:plot3d]{rgl::plot3d()}}. Not
used here.}
}
\value{
If \code{widget} is \code{TRUE}, an HTML widget is returned. Otherwise,
nothing.
}
\description{
Plots a 3D representation of a cylinder bundle from an object of class
\code{bundle} as generated by \code{\link[=simulate_bundle]{simulate_bundle()}} using \code{rgl}.
}
\examples{
density <- 0.9
voxel_size <- 0.000010 # 10 micrometers
out <- simulate_bundle(density, voxel_size)
rgl::plot3d(out)
}
