% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcylinders.R
\name{rcylinders}
\alias{rcylinders}
\title{Sample cylinder compartments}
\usage{
rcylinders(
  n,
  axis_mean,
  radius_mean,
  diffusivity_mean,
  axis_concentration = Inf,
  radius_sd = 0,
  diffusivity_sd = 0,
  restricted_model = c("callaghan", "neuman", "soderman", "stanisz", "vangelderen")
)
}
\arguments{
\item{n}{An integer value specifying the number of compartments to sample.}

\item{axis_mean}{A numeric value specifying the mean of the axis
distribution.}

\item{radius_mean}{A numeric value specifying the mean of the radius
distribution.}

\item{diffusivity_mean}{A numeric value specifying the mean of the
diffusivity distribution.}

\item{axis_concentration}{A numeric value specifying the concentration of the
axis distribution. Defaults to \code{Inf} in which case the axis is fixed to the
mean value.}

\item{radius_sd}{A numeric value specifying the standard deviation of the
radius distribution. Defaults to \code{0} in which case the radius is fixed to
the mean value.}

\item{diffusivity_sd}{A numeric value specifying the standard deviation of
the diffusivity distribution. Defaults to \code{0} in which case the diffusivity
is fixed to the mean value.}

\item{restricted_model}{A character vector specifying the restricted
diffusion model to use. Defaults to \code{callaghan}. Possible values are
\code{callaghan}, \code{neuman}, \code{soderman}, \code{stanisz}, and \code{vangelderen}.}
}
\value{
A list of \eqn{n} cylinder compartments of class
\code{\link{CylinderCompartment}}.
}
\description{
This function samples \eqn{n} cylinder compartments with given axis, radius
and diffusivity distributions.
}
\details{
The axis distribution is given by a mean and a concentration parameter and
the Dimroth-Watson distribution is used to sample values. The radius and
diffusivity distributions are given by a mean and a standard deviation and
the Gamma distribution is used to sample values. If the concentration
parameter is set to \code{Inf} the axis is fixed to the mean value. If the
standard deviation of the radius and diffusivity distributions are set to \code{0}
the radius and diffusivity are fixed to the mean values. If all parameters
are fixed, only one compartment is sampled.
}
\examples{
# Sample 10 cylinder compartments with fixed axis, radius and diffusivity
# set.seed(42)
cyl_distr <- rcylinders(
  n = 10L,
  axis_mean = c(0, 0, 1),
  radius_mean = 5,
  diffusivity_mean = 3
)
}
