% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-app.R
\name{run_app}
\alias{run_app}
\title{Runs the MIDI Shiny web application}
\usage{
run_app()
}
\value{
Nothing, but launches the Shiny app in the default web browser.
}
\description{
This is a helper function to run the MIDI Shiny web application in the
default web browser.
}
\examples{
\donttest{
  run_app()
}
}
