% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_scales.R
\name{scale_major}
\alias{scale_major}
\alias{scale_hungarian_minor}
\alias{scale_harmonic_minor}
\alias{scale_melodic_minor}
\alias{scale_minor}
\alias{scale_chromatic}
\title{scales}
\usage{
scale_major(key = "c", collapse = FALSE, ignore_octave = FALSE)

scale_hungarian_minor(key = "am", collapse = FALSE, ignore_octave = FALSE)

scale_harmonic_minor(key = "am", collapse = FALSE, ignore_octave = FALSE)

scale_melodic_minor(
  key = "am",
  descending = FALSE,
  collapse = FALSE,
  ignore_octave = FALSE
)

scale_minor(key = "am", collapse = FALSE, ignore_octave = FALSE)

scale_chromatic(
  root = "c",
  collapse = FALSE,
  sharp = TRUE,
  ignore_octave = FALSE
)
}
\arguments{
\item{key}{character, key signature.}

\item{collapse}{logical, collapse result into a single string ready for
phrase construction.}

\item{ignore_octave}{logical, strip octave numbering from scales not rooted
on C.}

\item{descending}{logical, return the descending scale, available as a
built-in argument for the melodic minor scale, which is different in each
direction.}

\item{root}{character, root note.}

\item{sharp}{logical, accidentals in arbitrary scale output should be sharp
rather than flat.}
}
\description{
Functions re-exported from tabr package.
}
