% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder-radial-compartment.R
\name{StaniszCompartment}
\alias{StaniszCompartment}
\title{Stanisz's model for restricted diffusion in a cylinder}
\description{
A class to model restricted diffusion in a cylinder using the
Stanisz's model.
}
\references{
Stanisz, G. J., Wright, G. A., Henkelman, R. M., & Szafer, A.
(1997). An analytical model of restricted diffusion in bovine optic nerve.
Magnetic Resonance in Medicine, 37(1), 103-111.
}
\section{Super class}{
\code{\link[midi:CylinderRadialCompartment]{midi::CylinderRadialCompartment}} -> \code{StaniszCompartment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StaniszCompartment-clone}{\code{StaniszCompartment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="get_signal"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-get_signal'><code>midi::CylinderRadialCompartment$get_signal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="initialize"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-initialize'><code>midi::CylinderRadialCompartment$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaniszCompartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StaniszCompartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaniszCompartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
