% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder-radial-compartment.R
\name{VanGelderenCompartment}
\alias{VanGelderenCompartment}
\title{Van Gelderen's model for restricted diffusion in a cylinder}
\description{
A class to model restricted diffusion in a cylinder using the
Van Gelderen's model.
}
\references{
Vangelderen, P., DesPres, D., Vanzijl, P. C. M., & Moonen, C. T.
W. (1994). Evaluation of restricted diffusion in cylinders. Phosphocreatine
in rabbit leg muscle. Journal of Magnetic Resonance, Series B, 103(3),
255-260.
}
\section{Super class}{
\code{\link[midi:CylinderRadialCompartment]{midi::CylinderRadialCompartment}} -> \code{VanGelderenCompartment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VanGelderenCompartment-clone}{\code{VanGelderenCompartment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="get_signal"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-get_signal'><code>midi::CylinderRadialCompartment$get_signal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="initialize"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-initialize'><code>midi::CylinderRadialCompartment$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VanGelderenCompartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VanGelderenCompartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VanGelderenCompartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
