% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{GammaDistribution}
\alias{GammaDistribution}
\title{Gamma distribution class}
\description{
This class defines the gamma distribution. It provides methods
for fitting the distribution to data and generating random samples.
}
\examples{
gd <- GammaDistribution$new(
  shape = 1,
  scale = 5
)
gd$get_shape()
gd$get_scale()
gd$get_mean()
gd$get_variance()
gd$random(10)
gd$fit(gd$random(100))
gd$get_shape()
gd$get_scale()
}
\section{Super class}{
\code{\link[midi:BaseDistribution]{midi::BaseDistribution}} -> \code{GammaDistribution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GammaDistribution-new}{\code{GammaDistribution$new()}}
\item \href{#method-GammaDistribution-get_shape}{\code{GammaDistribution$get_shape()}}
\item \href{#method-GammaDistribution-get_scale}{\code{GammaDistribution$get_scale()}}
\item \href{#method-GammaDistribution-clone}{\code{GammaDistribution$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="midi" data-topic="BaseDistribution" data-id="fit"><a href='../../midi/html/BaseDistribution.html#method-BaseDistribution-fit'><code>midi::BaseDistribution$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="BaseDistribution" data-id="get_mean"><a href='../../midi/html/BaseDistribution.html#method-BaseDistribution-get_mean'><code>midi::BaseDistribution$get_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="BaseDistribution" data-id="get_variance"><a href='../../midi/html/BaseDistribution.html#method-BaseDistribution-get_variance'><code>midi::BaseDistribution$get_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="BaseDistribution" data-id="random"><a href='../../midi/html/BaseDistribution.html#method-BaseDistribution-random'><code>midi::BaseDistribution$random()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-GammaDistribution-new}{}}}
\subsection{Method \code{new()}}{
Creates a new gamma distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaDistribution$new(shape = 1, scale = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{A numeric value specifying the shape parameter of the gamma
distribution. Defaults to \code{1.0}.}

\item{\code{scale}}{A numeric value specifying the scale parameter of the gamma
distribution. Defaults to \code{1.0}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the gamma distribution as an object of class
\code{\link{GammaDistribution}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaDistribution-get_shape"></a>}}
\if{latex}{\out{\hypertarget{method-GammaDistribution-get_shape}{}}}
\subsection{Method \code{get_shape()}}{
Retrieves the shape parameter of the gamma distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaDistribution$get_shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value storing the shape parameter of the gamma
distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaDistribution-get_scale"></a>}}
\if{latex}{\out{\hypertarget{method-GammaDistribution-get_scale}{}}}
\subsection{Method \code{get_scale()}}{
Retrieves the scale parameter of the gamma distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaDistribution$get_scale()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value storing the scale parameter of the gamma
distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GammaDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
