% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_R6.R
\name{format.midi_params}
\alias{format.midi_params}
\alias{vec_ptype_abbr.midi_params}
\title{Methods for displaying midi parameters in midi object output}
\usage{
\method{format}{midi_params}(x, ...)

vec_ptype_abbr.midi_params(x)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters passed to other methods}
}
\description{
Methods for displaying midi parameters in midi object output
}
