% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder-radial-compartment.R
\name{SodermanCompartment}
\alias{SodermanCompartment}
\title{Soderman's model for restricted diffusion in a cylinder}
\description{
A class to model restricted diffusion in a cylinder using the
Soderman's model.
}
\references{
Söderman, O., & Jönsson, B. (1995). Restricted diffusion in
cylindrical geometry. Journal of Magnetic Resonance, Series A, 117(1),
94-97.
}
\section{Super class}{
\code{\link[midi:CylinderRadialCompartment]{midi::CylinderRadialCompartment}} -> \code{SodermanCompartment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SodermanCompartment-clone}{\code{SodermanCompartment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="get_signal"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-get_signal'><code>midi::CylinderRadialCompartment$get_signal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="midi" data-topic="CylinderRadialCompartment" data-id="initialize"><a href='../../midi/html/CylinderRadialCompartment.html#method-CylinderRadialCompartment-initialize'><code>midi::CylinderRadialCompartment$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SodermanCompartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SodermanCompartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SodermanCompartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
