% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere-compartment.R
\name{SphereCompartment}
\alias{SphereCompartment}
\title{Sphere compartment class}
\description{
A class to model restricted diffusion in a sphere.
}
\examples{

## ------------------------------------------------
## Method `SphereCompartment$get_signal`
## ------------------------------------------------

sphereComp <- SphereCompartment$new(
  radius = 1e-6,
  diffusivity = 2.0e-9,
  model = "soderman"
)
sphereComp$get_signal(
  small_delta = 0.03,
  big_delta = 0.03,
  G = 0.040
)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SphereCompartment-new}{\code{SphereCompartment$new()}}
\item \href{#method-SphereCompartment-get_signal}{\code{SphereCompartment$get_signal()}}
\item \href{#method-SphereCompartment-clone}{\code{SphereCompartment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SphereCompartment-new"></a>}}
\if{latex}{\out{\hypertarget{method-SphereCompartment-new}{}}}
\subsection{Method \code{new()}}{
Instantiates a new sphere compartment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SphereCompartment$new(
  radius,
  diffusivity,
  model = c("soderman", "callaghan", "stanisz", "neuman", "vangelderen")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{radius}}{A numeric value specifying the radius of the sphere in
meters.}

\item{\code{diffusivity}}{A numeric value specifying the diffusivity within the
sphere in m\eqn{^2}.s\eqn{^{-1}}.}

\item{\code{model}}{A character string specifying the radial model to use.
Choices are \code{"soderman"}, \code{"callaghan"}, \code{"stanisz"}, \code{"neuman"}, and
\code{"vangelderen"}. Defaults to \code{"soderman"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{\link{SphereCompartment}} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SphereCompartment-get_signal"></a>}}
\if{latex}{\out{\hypertarget{method-SphereCompartment-get_signal}{}}}
\subsection{Method \code{get_signal()}}{
Computes the signal attenuation predicted by the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SphereCompartment$get_signal(
  small_delta,
  big_delta,
  G,
  echo_time = NULL,
  n_max = 20L,
  m_max = 50L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{small_delta}}{A numeric value specifying the duration of the
gradient pulse in seconds.}

\item{\code{big_delta}}{A numeric value specifying the duration between the
gradient pulses in seconds.}

\item{\code{G}}{A numeric value specifying the strength of the gradient in
T.m\eqn{^{-1}}.}

\item{\code{echo_time}}{A numeric value specifying the echo time in seconds.}

\item{\code{n_max}}{An integer value specifying the maximum order of the Bessel
function. Defaults to \code{20L}.}

\item{\code{m_max}}{An integer value specifying the maximum number of extrema
for the Bessel function. Defaults to \code{50L}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value storing the predicted signal attenuation.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{sphereComp <- SphereCompartment$new(
  radius = 1e-6,
  diffusivity = 2.0e-9,
  model = "soderman"
)
sphereComp$get_signal(
  small_delta = 0.03,
  big_delta = 0.03,
  G = 0.040
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SphereCompartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SphereCompartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SphereCompartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
